+++ tests/ServiceStack.Redis.Tests/Generic/RedisTypedTransactionTests.cs
@@ -121,7 +121,7 @@ namespace ServiceStack.Redis.Tests.Generic
 		public void Supports_different_operation_types_in_same_transaction()
 		{
 			var incrementResults = new List<long>();
-			var collectionCounts = new List<int>();
+			var collectionCounts = new List<long>();
 			var containsItem = false;
 
 			var typedList = typedClient.Lists[ListKey];
